fix(ci): python-coverage-shards — split oversized test files under 1500-line limit
This PR's new error-paths/github-client test files breached the maintainability
guard (test_first_party_code_files_stay_under_line_limit): part_004 of the
executor-error-paths suite hit 1663 lines and github_client part_004 hit 1504.
Relocate self-contained test classes without changing behavior — move
TestPrMonitorResume into a new error-paths part_019 (carrying only the helpers it
uses) and the create_pull_request URL-validation tests into a new class in
github_client part_003. No tests dropped; source coverage is unchanged.
