+++ tests/unit/common/test_github_client_parts/test_github_client_part_003.py
@@ -858,3 +858,54 @@ class TestFetchPrStatusPart002:
                 pr_number=1,
                 base_behind_count=0,
             )
+
+
+class TestCreatePullRequestUrlValidation:
+    """``create_pull_request`` stdout-URL extraction and validation.
+
+    Split out of ``test_github_client_part_004.py`` to keep each first-party test
+    file under the maintainability line limit.
+    """
+
+    @pytest.mark.unit
+    async def test_create_pull_request_extracts_url_from_noisy_stdout(self) -> None:
+        # gh may prefix the URL with status noise on stdout; we extract the URL
+        # rather than returning the whole line verbatim.
+        fake = FakeCommandRunner()
+        fake.queue_result(
+            returncode=0,
+            stdout=(
+                "Creating pull request for development into main in o/r\n"
+                "https://github.com/o/r/pull/777\n"
+            ),
+        )
+        client = GitHubClient(fake)
+        url = await client.create_pull_request(
+            repo=RepoRef(owner="o", name="r"),
+            base="main",
+            head="development",
+            title="t",
+            body="b",
+        )
+        assert url == "https://github.com/o/r/pull/777"
+
+    @pytest.mark.unit
+    async def test_create_pull_request_raises_when_stdout_has_no_url(self) -> None:
+        # gh exits 0 but prints only warning/status text — must raise rather than
+        # persist a non-URL string that breaks downstream PR-number extraction.
+        fake = FakeCommandRunner()
+        fake.queue_result(
+            returncode=0,
+            stdout="Warning: 1 uncommitted change\n",
+        )
+        client = GitHubClient(fake)
+        with pytest.raises(GitHubClientError) as exc:
+            await client.create_pull_request(
+                repo=RepoRef(owner="o", name="r"),
+                base="main",
+                head="development",
+                title="t",
+                body="b",
+            )
+        assert exc.value.operation == "gh pr create (no URL in stdout)"
+        assert "unexpected gh output" in str(exc.value)
